+++ src/util/mod.rs
@@ -2,6 +2,7 @@ mod accept;
 mod select;
 
 pub use accept::Accept;
+pub use accept::Bind;
 pub use accept::Listener;
 
 pub use select::select;